:sparkles: Renaming the application.
