+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.lastfmlevelone">
+  package="com.krrskl.lastfm">
 
     <uses-permission android:name="android.permission.INTERNET" />
