+++ android/app/src/main/java/com/krrskl/lastfm/MainApplication.java
@@ -1,4 +1,4 @@
-package com.lastfmlevelone;
+package com.krrskl.lastfm;
 
 import android.app.Application;
 import android.content.Context;
@@ -63,7 +63,7 @@ public class MainApplication extends NavigationApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.lastfmlevelone.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.krrskl.lastfm.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
