+++ android/app/src/main/res/raw/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "LastFMLevelOne",
+  "displayName": "LastFMLevelOne"
+}
