+++ android/app/src/main/res/raw/node_modules_reactnativevectoricons_glyphmaps_feather.json
@@ -1,0 +1,281 @@
+{
+  "activity": 61696,
+  "airplay": 61697,
+  "alert-circle": 61698,
+  "alert-octagon": 61699,
+  "alert-triangle": 61700,
+  "align-center": 61701,
+  "align-justify": 61702,
+  "align-left": 61703,
+  "align-right": 61704,
+  "anchor": 61705,
+  "aperture": 61706,
+  "archive": 61707,
+  "arrow-down": 61708,
+  "arrow-down-circle": 61709,
+  "arrow-down-left": 61710,
+  "arrow-down-right": 61711,
+  "arrow-left": 61712,
+  "arrow-left-circle": 61713,
+  "arrow-right": 61714,
+  "arrow-right-circle": 61715,
+  "arrow-up": 61716,
+  "arrow-up-circle": 61717,
+  "arrow-up-left": 61718,
+  "arrow-up-right": 61719,
+  "at-sign": 61720,
+  "award": 61721,
+  "bar-chart": 61722,
+  "bar-chart-2": 61723,
+  "battery": 61724,
+  "battery-charging": 61725,
+  "bell": 61726,
+  "bell-off": 61727,
+  "bluetooth": 61728,
+  "bold": 61729,
+  "book": 61730,
+  "book-open": 61731,
+  "bookmark": 61732,
+  "box": 61733,
+  "briefcase": 61734,
+  "calendar": 61735,
+  "camera": 61736,
+  "camera-off": 61737,
+  "cast": 61738,
+  "check": 61739,
+  "check-circle": 61740,
+  "check-square": 61741,
+  "chevron-down": 61742,
+  "chevron-left": 61743,
+  "chevron-right": 61744,
+  "chevron-up": 61745,
+  "chevrons-down": 61746,
+  "chevrons-left": 61747,
+  "chevrons-right": 61748,
+  "chevrons-up": 61749,
+  "chrome": 61750,
+  "circle": 61751,
+  "clipboard": 61752,
+  "clock": 61753,
+  "cloud": 61754,
+  "cloud-drizzle": 61755,
+  "cloud-lightning": 61756,
+  "cloud-off": 61757,
+  "cloud-rain": 61758,
+  "cloud-snow": 61759,
+  "code": 61760,
+  "codepen": 61761,
+  "codesandbox": 61972,
+  "coffee": 61762,
+  "columns": 61973,
+  "command": 61763,
+  "compass": 61764,
+  "copy": 61765,
+  "corner-down-left": 61766,
+  "corner-down-right": 61767,
+  "corner-left-down": 61768,
+  "corner-left-up": 61769,
+  "corner-right-down": 61770,
+  "corner-right-up": 61771,
+  "corner-up-left": 61772,
+  "corner-up-right": 61773,
+  "cpu": 61774,
+  "credit-card": 61775,
+  "crop": 61776,
+  "crosshair": 61777,
+  "database": 61778,
+  "delete": 61779,
+  "disc": 61780,
+  "dollar-sign": 61781,
+  "download": 61782,
+  "download-cloud": 61783,
+  "droplet": 61784,
+  "edit": 61785,
+  "edit-2": 61786,
+  "edit-3": 61787,
+  "external-link": 61788,
+  "eye": 61789,
+  "eye-off": 61790,
+  "facebook": 61791,
+  "fast-forward": 61792,
+  "feather": 61793,
+  "figma": 61970,
+  "file": 61794,
+  "file-minus": 61795,
+  "file-plus": 61796,
+  "file-text": 61797,
+  "film": 61798,
+  "filter": 61799,
+  "flag": 61800,
+  "folder": 61801,
+  "folder-minus": 61802,
+  "folder-plus": 61803,
+  "frown": 61804,
+  "gift": 61805,
+  "git-branch": 61806,
+  "git-commit": 61807,
+  "git-merge": 61808,
+  "git-pull-request": 61809,
+  "github": 61810,
+  "gitlab": 61811,
+  "globe": 61812,
+  "grid": 61813,
+  "hard-drive": 61814,
+  "hash": 61815,
+  "headphones": 61816,
+  "heart": 61817,
+  "help-circle": 61818,
+  "hexagon": 61974,
+  "home": 61819,
+  "image": 61820,
+  "inbox": 61821,
+  "info": 61822,
+  "instagram": 61823,
+  "italic": 61824,
+  "key": 61967,
+  "layers": 61825,
+  "layout": 61826,
+  "life-buoy": 61827,
+  "link": 61828,
+  "link-2": 61829,
+  "linkedin": 61830,
+  "list": 61831,
+  "loader": 61832,
+  "lock": 61833,
+  "log-in": 61834,
+  "log-out": 61835,
+  "mail": 61836,
+  "map": 61837,
+  "map-pin": 61838,
+  "maximize": 61839,
+  "maximize-2": 61840,
+  "meh": 61841,
+  "menu": 61842,
+  "message-circle": 61843,
+  "message-square": 61844,
+  "mic": 61845,
+  "mic-off": 61846,
+  "minimize": 61847,
+  "minimize-2": 61848,
+  "minus": 61849,
+  "minus-circle": 61850,
+  "minus-square": 61851,
+  "monitor": 61852,
+  "moon": 61853,
+  "more-horizontal": 61854,
+  "more-vertical": 61855,
+  "mouse-pointer": 61968,
+  "move": 61856,
+  "music": 61857,
+  "navigation": 61858,
+  "navigation-2": 61859,
+  "octagon": 61860,
+  "package": 61861,
+  "paperclip": 61862,
+  "pause": 61863,
+  "pause-circle": 61864,
+  "pen-tool": 61969,
+  "percent": 61865,
+  "phone": 61866,
+  "phone-call": 61867,
+  "phone-forwarded": 61868,
+  "phone-incoming": 61869,
+  "phone-missed": 61870,
+  "phone-off": 61871,
+  "phone-outgoing": 61872,
+  "pie-chart": 61873,
+  "play": 61874,
+  "play-circle": 61875,
+  "plus": 61876,
+  "plus-circle": 61877,
+  "plus-square": 61878,
+  "pocket": 61879,
+  "power": 61880,
+  "printer": 61881,
+  "radio": 61882,
+  "refresh-ccw": 61883,
+  "refresh-cw": 61884,
+  "repeat": 61885,
+  "rewind": 61886,
+  "rotate-ccw": 61887,
+  "rotate-cw": 61888,
+  "rss": 61889,
+  "save": 61890,
+  "scissors": 61891,
+  "search": 61892,
+  "send": 61893,
+  "server": 61894,
+  "settings": 61895,
+  "share": 61896,
+  "share-2": 61897,
+  "shield": 61898,
+  "shield-off": 61899,
+  "shopping-bag": 61900,
+  "shopping-cart": 61901,
+  "shuffle": 61902,
+  "sidebar": 61903,
+  "skip-back": 61904,
+  "skip-forward": 61905,
+  "slack": 61906,
+  "slash": 61907,
+  "sliders": 61908,
+  "smartphone": 61909,
+  "smile": 61910,
+  "speaker": 61911,
+  "square": 61912,
+  "star": 61913,
+  "stop-circle": 61914,
+  "sun": 61915,
+  "sunrise": 61916,
+  "sunset": 61917,
+  "tablet": 61975,
+  "tag": 61919,
+  "target": 61920,
+  "terminal": 61921,
+  "thermometer": 61922,
+  "thumbs-down": 61923,
+  "thumbs-up": 61924,
+  "toggle-left": 61925,
+  "toggle-right": 61926,
+  "trash": 61927,
+  "trash-2": 61928,
+  "trello": 61929,
+  "trending-down": 61930,
+  "trending-up": 61931,
+  "triangle": 61932,
+  "truck": 61933,
+  "tv": 61934,
+  "twitter": 61935,
+  "type": 61936,
+  "umbrella": 61937,
+  "underline": 61938,
+  "unlock": 61939,
+  "upload": 61940,
+  "upload-cloud": 61941,
+  "user": 61942,
+  "user-check": 61943,
+  "user-minus": 61944,
+  "user-plus": 61945,
+  "user-x": 61946,
+  "users": 61947,
+  "video": 61948,
+  "video-off": 61949,
+  "voicemail": 61950,
+  "volume": 61951,
+  "volume-1": 61952,
+  "volume-2": 61953,
+  "volume-x": 61954,
+  "watch": 61955,
+  "wifi": 61956,
+  "wifi-off": 61957,
+  "wind": 61958,
+  "x": 61959,
+  "x-circle": 61960,
+  "x-octagon": 61971,
+  "x-square": 61961,
+  "youtube": 61962,
+  "zap": 61963,
+  "zap-off": 61964,
+  "zoom-in": 61965,
+  "zoom-out": 61966
+}
