+++ android/app/src/main/res/raw/node_modules_reactnativevectoricons_glyphmaps_foundation.json
@@ -1,0 +1,285 @@
+{
+  "address-book": 61696,
+  "alert": 61697,
+  "align-center": 61698,
+  "align-justify": 61699,
+  "align-left": 61700,
+  "align-right": 61701,
+  "anchor": 61702,
+  "annotate": 61703,
+  "archive": 61704,
+  "arrow-down": 61705,
+  "arrow-left": 61706,
+  "arrow-right": 61707,
+  "arrow-up": 61708,
+  "arrows-compress": 61709,
+  "arrows-expand": 61710,
+  "arrows-in": 61711,
+  "arrows-out": 61712,
+  "asl": 61713,
+  "asterisk": 61714,
+  "at-sign": 61715,
+  "background-color": 61716,
+  "battery-empty": 61717,
+  "battery-full": 61718,
+  "battery-half": 61719,
+  "bitcoin-circle": 61720,
+  "bitcoin": 61721,
+  "blind": 61722,
+  "bluetooth": 61723,
+  "bold": 61724,
+  "book-bookmark": 61725,
+  "book": 61726,
+  "bookmark": 61727,
+  "braille": 61728,
+  "burst-new": 61729,
+  "burst-sale": 61730,
+  "burst": 61731,
+  "calendar": 61732,
+  "camera": 61733,
+  "check": 61734,
+  "checkbox": 61735,
+  "clipboard-notes": 61736,
+  "clipboard-pencil": 61737,
+  "clipboard": 61738,
+  "clock": 61739,
+  "closed-caption": 61740,
+  "cloud": 61741,
+  "comment-minus": 61742,
+  "comment-quotes": 61743,
+  "comment-video": 61744,
+  "comment": 61745,
+  "comments": 61746,
+  "compass": 61747,
+  "contrast": 61748,
+  "credit-card": 61749,
+  "crop": 61750,
+  "crown": 61751,
+  "css3": 61752,
+  "database": 61753,
+  "die-five": 61754,
+  "die-four": 61755,
+  "die-one": 61756,
+  "die-six": 61757,
+  "die-three": 61758,
+  "die-two": 61759,
+  "dislike": 61760,
+  "dollar-bill": 61761,
+  "dollar": 61762,
+  "download": 61763,
+  "eject": 61764,
+  "elevator": 61765,
+  "euro": 61766,
+  "eye": 61767,
+  "fast-forward": 61768,
+  "female-symbol": 61769,
+  "female": 61770,
+  "filter": 61771,
+  "first-aid": 61772,
+  "flag": 61773,
+  "folder-add": 61774,
+  "folder-lock": 61775,
+  "folder": 61776,
+  "foot": 61777,
+  "foundation": 61778,
+  "graph-bar": 61779,
+  "graph-horizontal": 61780,
+  "graph-pie": 61781,
+  "graph-trend": 61782,
+  "guide-dog": 61783,
+  "hearing-aid": 61784,
+  "heart": 61785,
+  "home": 61786,
+  "html5": 61787,
+  "indent-less": 61788,
+  "indent-more": 61789,
+  "info": 61790,
+  "italic": 61791,
+  "key": 61792,
+  "laptop": 61793,
+  "layout": 61794,
+  "lightbulb": 61795,
+  "like": 61796,
+  "link": 61797,
+  "list-bullet": 61798,
+  "list-number": 61799,
+  "list-thumbnails": 61800,
+  "list": 61801,
+  "lock": 61802,
+  "loop": 61803,
+  "magnifying-glass": 61804,
+  "mail": 61805,
+  "male-female": 61806,
+  "male-symbol": 61807,
+  "male": 61808,
+  "map": 61809,
+  "marker": 61810,
+  "megaphone": 61811,
+  "microphone": 61812,
+  "minus-circle": 61813,
+  "minus": 61814,
+  "mobile-signal": 61815,
+  "mobile": 61816,
+  "monitor": 61817,
+  "mountains": 61818,
+  "music": 61819,
+  "next": 61820,
+  "no-dogs": 61821,
+  "no-smoking": 61822,
+  "page-add": 61823,
+  "page-copy": 61824,
+  "page-csv": 61825,
+  "page-delete": 61826,
+  "page-doc": 61827,
+  "page-edit": 61828,
+  "page-export-csv": 61829,
+  "page-export-doc": 61830,
+  "page-export-pdf": 61831,
+  "page-export": 61832,
+  "page-filled": 61833,
+  "page-multiple": 61834,
+  "page-pdf": 61835,
+  "page-remove": 61836,
+  "page-search": 61837,
+  "page": 61838,
+  "paint-bucket": 61839,
+  "paperclip": 61840,
+  "pause": 61841,
+  "paw": 61842,
+  "paypal": 61843,
+  "pencil": 61844,
+  "photo": 61845,
+  "play-circle": 61846,
+  "play-video": 61847,
+  "play": 61848,
+  "plus": 61849,
+  "pound": 61850,
+  "power": 61851,
+  "previous": 61852,
+  "price-tag": 61853,
+  "pricetag-multiple": 61854,
+  "print": 61855,
+  "prohibited": 61856,
+  "projection-screen": 61857,
+  "puzzle": 61858,
+  "quote": 61859,
+  "record": 61860,
+  "refresh": 61861,
+  "results-demographics": 61862,
+  "results": 61863,
+  "rewind-ten": 61864,
+  "rewind": 61865,
+  "rss": 61866,
+  "safety-cone": 61867,
+  "save": 61868,
+  "share": 61869,
+  "sheriff-badge": 61870,
+  "shield": 61871,
+  "shopping-bag": 61872,
+  "shopping-cart": 61873,
+  "shuffle": 61874,
+  "skull": 61875,
+  "social-500px": 61876,
+  "social-adobe": 61877,
+  "social-amazon": 61878,
+  "social-android": 61879,
+  "social-apple": 61880,
+  "social-behance": 61881,
+  "social-bing": 61882,
+  "social-blogger": 61883,
+  "social-delicious": 61884,
+  "social-designer-news": 61885,
+  "social-deviant-art": 61886,
+  "social-digg": 61887,
+  "social-dribbble": 61888,
+  "social-drive": 61889,
+  "social-dropbox": 61890,
+  "social-evernote": 61891,
+  "social-facebook": 61892,
+  "social-flickr": 61893,
+  "social-forrst": 61894,
+  "social-foursquare": 61895,
+  "social-game-center": 61896,
+  "social-github": 61897,
+  "social-google-plus": 61898,
+  "social-hacker-news": 61899,
+  "social-hi5": 61900,
+  "social-instagram": 61901,
+  "social-joomla": 61902,
+  "social-lastfm": 61903,
+  "social-linkedin": 61904,
+  "social-medium": 61905,
+  "social-myspace": 61906,
+  "social-orkut": 61907,
+  "social-path": 61908,
+  "social-picasa": 61909,
+  "social-pinterest": 61910,
+  "social-rdio": 61911,
+  "social-reddit": 61912,
+  "social-skillshare": 61913,
+  "social-skype": 61914,
+  "social-smashing-mag": 61915,
+  "social-snapchat": 61916,
+  "social-spotify": 61917,
+  "social-squidoo": 61918,
+  "social-stack-overflow": 61919,
+  "social-steam": 61920,
+  "social-stumbleupon": 61921,
+  "social-treehouse": 61922,
+  "social-tumblr": 61923,
+  "social-twitter": 61924,
+  "social-vimeo": 61925,
+  "social-windows": 61926,
+  "social-xbox": 61927,
+  "social-yahoo": 61928,
+  "social-yelp": 61929,
+  "social-youtube": 61930,
+  "social-zerply": 61931,
+  "social-zurb": 61932,
+  "sound": 61933,
+  "star": 61934,
+  "stop": 61935,
+  "strikethrough": 61936,
+  "subscript": 61937,
+  "superscript": 61938,
+  "tablet-landscape": 61939,
+  "tablet-portrait": 61940,
+  "target-two": 61941,
+  "target": 61942,
+  "telephone-accessible": 61943,
+  "telephone": 61944,
+  "text-color": 61945,
+  "thumbnails": 61946,
+  "ticket": 61947,
+  "torso-business": 61948,
+  "torso-female": 61949,
+  "torso": 61950,
+  "torsos-all-female": 61951,
+  "torsos-all": 61952,
+  "torsos-female-male": 61953,
+  "torsos-male-female": 61954,
+  "torsos": 61955,
+  "trash": 61956,
+  "trees": 61957,
+  "trophy": 61958,
+  "underline": 61959,
+  "universal-access": 61960,
+  "unlink": 61961,
+  "unlock": 61962,
+  "upload-cloud": 61963,
+  "upload": 61964,
+  "usb": 61965,
+  "video": 61966,
+  "volume-none": 61967,
+  "volume-strike": 61968,
+  "volume": 61969,
+  "web": 61970,
+  "wheelchair": 61971,
+  "widget": 61972,
+  "wrench": 61973,
+  "x-circle": 61974,
+  "x": 61975,
+  "yen": 61976,
+  "zoom-in": 61977,
+  "zoom-out": 61978
+}
