+++ android/app/src/main/res/raw/node_modules_reactnativevectoricons_glyphmaps_octicons.json
@@ -1,0 +1,186 @@
+{
+  "alert": 61696,
+  "archive": 61697,
+  "arrow-both": 61698,
+  "arrow-down": 61699,
+  "arrow-left": 61700,
+  "arrow-right": 61701,
+  "arrow-small-down": 61702,
+  "arrow-small-left": 61703,
+  "arrow-small-right": 61704,
+  "arrow-small-up": 61705,
+  "arrow-up": 61706,
+  "beaker": 61707,
+  "bell": 61708,
+  "bold": 61709,
+  "book": 61710,
+  "bookmark": 61711,
+  "briefcase": 61712,
+  "broadcast": 61713,
+  "browser": 61714,
+  "bug": 61715,
+  "calendar": 61716,
+  "check": 61717,
+  "checklist": 61718,
+  "chevron-down": 61719,
+  "chevron-left": 61720,
+  "chevron-right": 61721,
+  "chevron-up": 61722,
+  "circle-slash": 61723,
+  "circuit-board": 61724,
+  "clippy": 61725,
+  "clock": 61726,
+  "cloud-download": 61727,
+  "cloud-upload": 61728,
+  "code": 61729,
+  "comment": 61730,
+  "comment-discussion": 61731,
+  "credit-card": 61732,
+  "dash": 61733,
+  "dashboard": 61734,
+  "database": 61735,
+  "desktop-download": 61736,
+  "device-camera": 61737,
+  "device-camera-video": 61738,
+  "device-desktop": 61739,
+  "device-mobile": 61740,
+  "diff": 61741,
+  "diff-added": 61742,
+  "diff-ignored": 61743,
+  "diff-modified": 61744,
+  "diff-removed": 61745,
+  "diff-renamed": 61746,
+  "ellipsis": 61747,
+  "eye": 61748,
+  "eye-closed": 61749,
+  "file": 61750,
+  "file-binary": 61751,
+  "file-code": 61752,
+  "file-directory": 61753,
+  "file-media": 61754,
+  "file-pdf": 61755,
+  "file-submodule": 61756,
+  "file-symlink-directory": 61757,
+  "file-symlink-file": 61758,
+  "file-zip": 61759,
+  "flame": 61760,
+  "fold": 61761,
+  "fold-down": 61762,
+  "fold-up": 61763,
+  "gear": 61764,
+  "gift": 61765,
+  "gist": 61766,
+  "gist-secret": 61767,
+  "git-branch": 61768,
+  "git-commit": 61769,
+  "git-compare": 61770,
+  "git-merge": 61771,
+  "git-pull-request": 61772,
+  "github-action": 61773,
+  "globe": 61774,
+  "grabber": 61775,
+  "graph": 61776,
+  "heart": 61777,
+  "history": 61778,
+  "home": 61779,
+  "horizontal-rule": 61780,
+  "hubot": 61781,
+  "inbox": 61782,
+  "info": 61783,
+  "issue-closed": 61784,
+  "issue-opened": 61785,
+  "issue-reopened": 61786,
+  "italic": 61787,
+  "jersey": 61788,
+  "kebab-horizontal": 61789,
+  "kebab-vertical": 61790,
+  "key": 61791,
+  "keyboard": 61792,
+  "law": 61793,
+  "light-bulb": 61794,
+  "link": 61795,
+  "link-external": 61796,
+  "list-ordered": 61797,
+  "list-unordered": 61798,
+  "location": 61799,
+  "lock": 61800,
+  "logo-gist": 61801,
+  "logo-github": 61802,
+  "mail": 61803,
+  "mail-read": 61804,
+  "mark-github": 61805,
+  "markdown": 61806,
+  "megaphone": 61807,
+  "mention": 61808,
+  "milestone": 61809,
+  "mirror": 61810,
+  "mortar-board": 61811,
+  "mute": 61812,
+  "no-newline": 61813,
+  "note": 61814,
+  "octoface": 61815,
+  "organization": 61816,
+  "package": 61817,
+  "paintcan": 61818,
+  "pencil": 61819,
+  "person": 61820,
+  "pin": 61821,
+  "play": 61822,
+  "plug": 61823,
+  "plus": 61824,
+  "plus-small": 61825,
+  "primitive-dot": 61826,
+  "primitive-square": 61827,
+  "project": 61828,
+  "pulse": 61829,
+  "question": 61830,
+  "quote": 61831,
+  "radio-tower": 61832,
+  "reply": 61833,
+  "repo": 61834,
+  "repo-clone": 61835,
+  "repo-force-push": 61836,
+  "repo-forked": 61837,
+  "repo-pull": 61838,
+  "repo-push": 61839,
+  "report": 61840,
+  "request-changes": 61841,
+  "rocket": 61842,
+  "rss": 61843,
+  "ruby": 61844,
+  "screen-full": 61845,
+  "screen-normal": 61846,
+  "search": 61847,
+  "server": 61848,
+  "settings": 61849,
+  "shield": 61850,
+  "sign-in": 61851,
+  "sign-out": 61852,
+  "smiley": 61853,
+  "squirrel": 61854,
+  "star": 61855,
+  "stop": 61856,
+  "sync": 61857,
+  "tag": 61858,
+  "tasklist": 61859,
+  "telescope": 61860,
+  "terminal": 61861,
+  "text-size": 61862,
+  "three-bars": 61863,
+  "thumbsdown": 61864,
+  "thumbsup": 61865,
+  "tools": 61866,
+  "trashcan": 61867,
+  "triangle-down": 61868,
+  "triangle-left": 61869,
+  "triangle-right": 61870,
+  "triangle-up": 61871,
+  "unfold": 61872,
+  "unmute": 61873,
+  "unverified": 61874,
+  "verified": 61875,
+  "versions": 61876,
+  "watch": 61877,
+  "x": 61878,
+  "zap": 61879
+}
