+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "LastFMLevelOne",
-  "displayName": "LastFMLevelOne"
+  "name": "LastFMLevelTwoAndThree",
+  "displayName": "LastFMLevelTwoAndThree"
 }
