+++ ios/LastFMLevelTwoAndThree.xcodeproj/xcshareddata/xcschemes/LastFMLevelTwoAndThree.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "LastFMLevelOne.app"
-               BlueprintName = "LastFMLevelOne"
-               ReferencedContainer = "container:LastFMLevelOne.xcodeproj">
+               BuildableName = "LastFMLevelTwoAndThree.app"
+               BlueprintName = "LastFMLevelTwoAndThree"
+               ReferencedContainer = "container:LastFMLevelTwoAndThree.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "LastFMLevelOneTests.xctest"
-               BlueprintName = "LastFMLevelOneTests"
-               ReferencedContainer = "container:LastFMLevelOne.xcodeproj">
+               BuildableName = "LastFMLevelTwoAndThreeTests.xctest"
+               BlueprintName = "LastFMLevelTwoAndThreeTests"
+               ReferencedContainer = "container:LastFMLevelTwoAndThree.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "LastFMLevelOne.app"
-            BlueprintName = "LastFMLevelOne"
-            ReferencedContainer = "container:LastFMLevelOne.xcodeproj">
+            BuildableName = "LastFMLevelTwoAndThree.app"
+            BlueprintName = "LastFMLevelTwoAndThree"
+            ReferencedContainer = "container:LastFMLevelTwoAndThree.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "LastFMLevelOne.app"
-            BlueprintName = "LastFMLevelOne"
-            ReferencedContainer = "container:LastFMLevelOne.xcodeproj">
+            BuildableName = "LastFMLevelTwoAndThree.app"
+            BlueprintName = "LastFMLevelTwoAndThree"
+            ReferencedContainer = "container:LastFMLevelTwoAndThree.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
