+++ ios/LastFMLevelTwoAndThree/Images.xcassets/AppIcon.appiconset/Contents.json
