+++ ios/LastFMLevelTwoAndThreeTests/LastFMLevelTwoAndThreeTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface LastFMLevelOneTests : XCTestCase
+@interface LastFMLevelTwoAndThreeTests : XCTestCase
 
 @end
 
-@implementation LastFMLevelOneTests
+@implementation LastFMLevelTwoAndThreeTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
