+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "LastFMLevelOne",
+  "name": "LastFMLevelTwoAndThree",
   "version": "0.0.1",
   "private": true,
   "scripts": {
